+++ src/agent0/core/hyperdrive/utilities/run_bots/__init__.py
@@ -1,7 +1,7 @@
 """Functions for running test experiment loops."""
 
 from .create_and_fund_user_account import create_and_fund_user_account
-from .execute_agent_trades import async_execute_multi_agent_trades
+from .execute_multi_agent_trades import async_execute_multi_agent_trades
 from .fund_agents import async_fund_agents
 from .run_agents import setup_and_run_agent_loop
 from .setup_experiment import setup_experiment
